backend/requirements.txt
torch
transformers
fastapi
elasticsearch==7.10.0
groq
faiss-cpu
python-dotenv
sentence-transformers
chromadb
